Fix int conversion
git-svn-id: file:///home/svn/DVA2006@15216 1e6d2d28-e1fb-0310-b8ab-81908d8c6b6a

+++ src/java/org/deltava/service/stats/SimulatorStatsService.java
@@ -15,7 +15,7 @@ import org.deltava.service.*;
 /**
  * A Web Service to graph simulator version statistics.
  * @author Luke
- * @version 9.0
+ * @version 9.1
  * @since 7.4
  */
 
@@ -50,7 +50,7 @@ public class SimulatorStatsService extends WebService {
 			ma.put(legs.getOrDefault("FSX", ZERO).intValue());
 			ma.put(legs.getOrDefault("FS9", ZERO).intValue());
 			ma.put(legs.getOrDefault("XP10", ZERO).intValue() + legs.getOrDefault("XP11", ZERO).intValue());
-			ma.put(legs.getOrDefault("FS2020", ZERO));
+			ma.put(legs.getOrDefault("FS2020", ZERO).intValue());
 			ma.put(legs.getOrDefault("FS2002", ZERO).intValue());
 			ma.put(legs.getOrDefault("UNKNOWN", ZERO).intValue() + legs.getOrDefault("FS2000", ZERO).intValue());
 			ja.put(ma);
